services/vuelos_service/requirements.txt
django
djangorestframework
djangorestframework-authtoken
requests
mysqlclient
gunicorn